feat(molitao_app): 完善登录流程和个人中心功能
- 应用启动时使用 token 自动获取用户信息
- 登录成功后调用 getCurrentLoginInformations 获取用户信息
- 用户未登录时不连接 WebSocket
- 修复好友列表 status 参数类型 (int -> bool)
- 个人中心页面每次进入重新请求好友数和魔力值
- 修复 depositBalance 类型 (int -> double)
- 退出登录按钮居中布局并添加确认对话框
- 添加调试日志便于问题排查

+++ molitao_app/lib/domain/entities/my_count_entity.dart
@@ -1,13 +1,13 @@
 class MyCountEntity {
   final int friend;
-  final int depositBalance;
+  final double depositBalance;
 
   MyCountEntity({required this.friend, required this.depositBalance});
 
   factory MyCountEntity.fromJson(Map<String, dynamic> json) {
     return MyCountEntity(
       friend: json['friend'] ?? 0,
-      depositBalance: json['depositBalance'] ?? 0,
+      depositBalance: (json['depositBalance'] ?? 0).toDouble(),
     );
   }
